Regroupe les classes relatives aux situations.
