--- app/models/competence/base.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Competence
-  class Base
-    def initialize(evaluation)
-      @evaluation = evaluation
-    end
-
-    def niveau
-      raise NotImplementedError
-    end
-  end
-end
