+++ spec/models/evaluation/controle/attention_concentration_spec.rb
@@ -2,11 +2,12 @@
 
 require 'rails_helper'
 
-describe Competence::ControleRapidite do
+describe Evaluation::Controle::AttentionConcentration do
   let(:evaluation) { double }
 
-  context "lorsqu'il n'y a pas de ratées" do
+  context "lorsqu'il n'y a pas d'erreurs ou de ratées" do
     it 'a le niveau 4' do
+      expect(evaluation).to receive(:nombre_mal_placees).and_return(0)
       expect(evaluation).to receive(:nombre_ratees).and_return(0)
       expect(
         described_class.new(evaluation).niveau
@@ -14,8 +15,19 @@ describe Competence::ControleRapidite do
     end
   end
 
+  context "lorsqu'il y a une pièce mal placée" do
+    it 'a le niveau 3' do
+      expect(evaluation).to receive(:nombre_mal_placees).and_return(1)
+      expect(evaluation).to receive(:nombre_ratees).and_return(0)
+      expect(
+        described_class.new(evaluation).niveau
+      ).to eql(Competence::NIVEAU_3)
+    end
+  end
+
   context "lorsqu'il y a une pièce ratée" do
     it 'a le niveau 3' do
+      expect(evaluation).to receive(:nombre_mal_placees).and_return(0)
       expect(evaluation).to receive(:nombre_ratees).and_return(1)
       expect(
         described_class.new(evaluation).niveau
@@ -25,6 +37,7 @@ describe Competence::ControleRapidite do
 
   context "lorsqu'il y a deux pièces ratées" do
     it 'a le niveau 2' do
+      expect(evaluation).to receive(:nombre_mal_placees).and_return(0)
       expect(evaluation).to receive(:nombre_ratees).and_return(2)
       expect(
         described_class.new(evaluation).niveau
@@ -34,6 +47,17 @@ describe Competence::ControleRapidite do
 
   context "lorsqu'il y a trois pièces ratées" do
     it 'a le niveau 1' do
+      expect(evaluation).to receive(:nombre_mal_placees).and_return(0)
+      expect(evaluation).to receive(:nombre_ratees).and_return(3)
+      expect(
+        described_class.new(evaluation).niveau
+      ).to eql(Competence::NIVEAU_1)
+    end
+  end
+
+  context "lorsqu'il y a quatre pièces ratées" do
+    it 'a le niveau 1' do
+      expect(evaluation).to receive(:nombre_mal_placees).and_return(1)
       expect(evaluation).to receive(:nombre_ratees).and_return(3)
       expect(
         described_class.new(evaluation).niveau
